requirements.txt
asgiref==3.7.2
Django==4.2.3
pip==23.1.2
setuptools==67.8.0
sqlparse==0.4.4
typing-extensions==4.7.1
wheel==0.40.0


djangorestframework
markdown
django-filter

# needed for the database
psycopg2-binary

# for the PSTDateTime
python-dateutil
pytz